Movie_Recommendation_System/requirements.txt
psycopg2-binary
requests
redis
gunicorn
whitenoise
django-redis
drf-yasg
psycopg2-binary
python-dotenv
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-redis
django-filter
dj-database-url
django-phonenumber-field
django-phonenumber-field[phonenumbers]
django-phonenumber-field[phonenumberslite]